PERF: Reduce filesize of `component/emoji-group-sections`
Why this change?

When using the `webpack-bundle-analyzer` plugin, I noticed that
`emoji-group-sections.hbs` accounts for 347KB and is significantly
bigger than most components. In turns out the file size is quite big
because we keep repeating the `{{replace-emoji ...}` line in the
handlebars file.

What does this change do?

Instead of repeating the `{{replace-emoji ...}` lines which inflates the
file size. We can instead just define the emojis in a constant and then
just loop through the elements in the constant to render the HTML
elements.

This reduces the parsed size from about 347KB to 38KB.